+++ backend/src/main/java/com/example/weight_inspection/controllers/PaletteController.java
@@ -10,7 +10,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
-import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
@@ -25,7 +24,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.example.weight_inspection.models.Palette;
-import com.example.weight_inspection.payloads.ListResponse;
+import com.example.weight_inspection.transfer.ListResponse;
 import com.example.weight_inspection.repositories.PaletteRepository;
 
 @RestController
